# features/configuration/dto/responses/pin.go
package responses

type ValidatePinResponse struct {
	Message string `json:"message"`
} // @name ValidatePinResponse

type ChangePinResponse struct {
	Message string `json:"message"`
} // @name ChangePinResponse